Add basic Moesif support for BI runtimes
