+++ icp_server/resources/db/init-scripts/h2_init.sql
@@ -121,6 +121,21 @@ CREATE TABLE components (
 
 CREATE INDEX idx_components_project_id ON components (project_id);
 
+-- Per-integration Moesif metrics configuration (optional feature). Kept in a
+-- separate table so the credential (management_key) and provider-specific
+-- settings do not widen the core components table. One row per configured
+-- component; removed automatically when the component is deleted.
+CREATE TABLE component_moesif_config (
+    component_id CHAR(36) PRIMARY KEY,
+    application_id VARCHAR(512),
+    dashboards_created BOOLEAN NOT NULL DEFAULT FALSE,
+    workspace_id VARCHAR(512),
+    management_key VARCHAR(4096),
+    created_at TIMESTAMP NOT NULL DEFAULT CURRENT_TIMESTAMP,
+    updated_at TIMESTAMP NOT NULL DEFAULT CURRENT_TIMESTAMP,
+    CONSTRAINT fk_moesif_config_component FOREIGN KEY (component_id) REFERENCES components (component_id) ON DELETE CASCADE
+);
+
 CREATE TABLE environments (
     environment_id CHAR(36) PRIMARY KEY,
     name VARCHAR(200) NOT NULL UNIQUE,
