+++ icp_server/resources/db/init-scripts/oracle_init.sql
@@ -172,6 +172,27 @@ BEGIN
 END;
 /
 
+-- Moesif metrics configuration for a component (kept separate from components to
+-- isolate provider secrets and avoid sparse columns on the core table).
+CREATE TABLE component_moesif_config (
+    component_id CHAR(36) PRIMARY KEY,
+    application_id VARCHAR2(512 CHAR) NULL,
+    dashboards_created NUMBER(1) DEFAULT 0 NOT NULL CHECK (dashboards_created IN (0, 1)),
+    workspace_id VARCHAR2(512 CHAR) NULL,
+    management_key VARCHAR2(4000 CHAR) NULL,
+    created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP NOT NULL,
+    updated_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP NOT NULL,
+    CONSTRAINT fk_moesif_config_component FOREIGN KEY (component_id) REFERENCES components (component_id) ON DELETE CASCADE
+);
+
+CREATE OR REPLACE TRIGGER trg_moesif_config_updated_at
+BEFORE UPDATE ON component_moesif_config
+FOR EACH ROW
+BEGIN
+    :NEW.updated_at := CURRENT_TIMESTAMP;
+END;
+/
+
 CREATE TABLE environments (
     environment_id CHAR(36) PRIMARY KEY,
     name VARCHAR2(200 CHAR) NOT NULL UNIQUE, -- e.g., dev, stage, prod
