+++ icp_server/resources/db/init-scripts/postgresql_init.sql
@@ -142,6 +142,22 @@ CREATE INDEX idx_comp_project_id ON components(project_id);
 CREATE TRIGGER update_components_updated_at BEFORE UPDATE ON components
     FOR EACH ROW EXECUTE FUNCTION update_updated_at_column();
 
+-- Moesif metrics configuration for a component (kept separate from components to
+-- isolate provider secrets and avoid sparse columns on the core table).
+CREATE TABLE component_moesif_config (
+    component_id CHAR(36) PRIMARY KEY,
+    application_id VARCHAR(512) NULL,
+    dashboards_created BOOLEAN NOT NULL DEFAULT FALSE,
+    workspace_id VARCHAR(512) NULL,
+    management_key VARCHAR(4096) NULL,
+    created_at TIMESTAMP NOT NULL DEFAULT CURRENT_TIMESTAMP,
+    updated_at TIMESTAMP NOT NULL DEFAULT CURRENT_TIMESTAMP,
+    CONSTRAINT fk_moesif_config_component FOREIGN KEY (component_id) REFERENCES components (component_id) ON DELETE CASCADE
+);
+
+CREATE TRIGGER update_component_moesif_config_updated_at BEFORE UPDATE ON component_moesif_config
+    FOR EACH ROW EXECUTE FUNCTION update_updated_at_column();
+
 CREATE TABLE environments (
     environment_id CHAR(36) PRIMARY KEY,
     name VARCHAR(200) NOT NULL UNIQUE, -- e.g., dev, stage, prod
